+++ ambra_sdk/async_addon/job.py
@@ -114,14 +114,23 @@ class Job:
                     break
                 with suppress(NotFound):
                     job_status = cast(Box, await get_job_query.get())
-                    if job_status['state'] != 'DONE':
+                    if job_status['state'] not in (  # NOQA: WPS510
+                        'FINISHED',
+                        'DONE',
+                        'ERROR',
+                        'RUNNING',
+                        'SCHEDULED',
+                    ):
                         raise RuntimeError(  # NOQA:WPS220
                             'Unknown job status {job_status}'.format(
                                 job_status=job_status['state'],
                             ),
                         )
-                    job_is_ready = True
-                    break
+                    if job_status['state'] == 'DONE':
+                        job_is_ready = True
+                        break
+                    elif job_status['state'] == 'ERROR':
+                        raise RuntimeError(f'Job finished with error {job_status}')
                 with suppress(TimeoutError):
                     await ws.wait_for_event(
                         channel_name,