+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [3.22.3.0-1] - 2022-05-25
+### Added
+- Service public api support: LBL0022 v53.0 2021-04-05
+- Storage public api support: LBL0038 v22.0 2022-05-18
+
 ## [3.22.1.0-1] - 2022-04-12
 ### Added
 - Service public api support: LBL0022 v53.0 2021-04-05